update int --to--long long

+++ CSES/increasing_Array.cpp
@@ -3,17 +3,17 @@ using namespace std;
 
 int main()
 {
-    int n;
+    long long n;
     cin>>n;
 
-    vector<int> arr(n);
-    for(int i=0; i<n; i++)
+    vector<long long> arr(n);
+    for(long long i=0; i<n; i++)
     {
         cin>>arr[i];
     }
 
     int moves = 0;
-    for(int i=1; i<n; i++)
+    for(long long i=1; i<n; i++)
     {
         if(arr[i]<arr[i-1])
         {
